+++ app/src/Fisica/Colisiones.hs
@@ -1,7 +1,7 @@
 module Fisica.Colisiones where
 
 import Linear.V2 (V2(..))
-import Types (Jugador(..), Enemigo(..), Obstaculo(..))
+import Types (Jugador(..), Enemigo(..), Obstaculo(..), Item(..))
 
 class Hitbox a where
     getPos :: a -> V2 Float
@@ -15,6 +15,10 @@ instance Hitbox Enemigo where
     getPos = posEnemigo
     getTam = tamEnemigo
 
+instance Hitbox Item where
+    getPos = posItem
+    getTam = tamItem
+
 -- Verifica si dos rectángulos (pos1, size1) y (pos2, size2) se superponen
 haySolapamiento :: V2 Float -> V2 Float -> V2 Float -> V2 Float -> Bool
 haySolapamiento (V2 x1 y1) (V2 w1 h1) (V2 x2 y2) (V2 w2 h2) =
@@ -23,7 +27,7 @@ haySolapamiento (V2 x1 y1) (V2 w1 h1) (V2 x2 y2) (V2 w2 h2) =
     y1 < y2 + h2 &&
     y1 + h1 > y2
 
--- Verifica si el jugador choca contra algún obstáculo de la lista
+-- Verifica si algo con Hitbox choca contra algún obstáculo de la lista
 checkColision :: Hitbox a => a -> [Obstaculo] -> Bool
 checkColision entidad obstaculos =
     any chocaCon obstaculos
@@ -32,3 +36,16 @@ checkColision entidad obstaculos =
     tamE = getTam entidad
     chocaCon (Obstaculo posObj tamObj) =
         haySolapamiento posE tamE posObj tamObj
+
+-- Verifica si algo con Hitbox choca contra algún item de la lista
+checkColisionsItems :: Hitbox a => a -> [Item] -> [Item]
+checkColisionsItems entidad listaItems =
+    filter estaTocando listaItems
+  where
+    posE = getPos entidad
+    tamE = getTam entidad
+    
+    estaTocando :: Item -> Bool
+    estaTocando it = 
+        -- Solo detectamos colisión si el item está "activo"
+        haySolapamiento posE tamE (posItem it) (tamItem it)