+++ app/src/Fisica/MovEnemigo.hs
@@ -1,11 +1,17 @@
 module Fisica.MovEnemigo where
 
 import Linear.V2 (V2(..))
-import Linear.Metric (normalize, distance)
+import Linear.Metric (normalize, distance, norm)
 import Linear.Vector ((^*))
 
 import Types
 import Fisica.Colisiones (checkColision)
+
+friccion :: Float
+friccion = 0.90
+
+umbralParada :: Float
+umbralParada = 0.5
 
 calcularDirEnemigo :: Enemigo -> Jugador -> V2 Float
 calcularDirEnemigo enemigoIni player =
@@ -20,10 +26,14 @@ calcularDirEnemigo enemigoIni player =
            V2 0 0
 
 -- Mover al enemigo
-moverEnemigo :: Enemigo -> V2 Float -> Mapa -> Enemigo
+moverEnemigo :: Enemigo -> V2 Float -> [Obstaculo] -> Enemigo
 moverEnemigo enemigoIni delta mapObstaculos = 
     let 
-        (V2 dx dy) = delta
+        esEmpujado = norm (velGolpeE enemigoIni) > umbralParada
+        
+        (V2 dx dy) = if esEmpujado 
+                     then velGolpeE enemigoIni
+                     else delta
         currentPos = posEnemigo enemigoIni
         
         -- Mover en X
@@ -37,7 +47,13 @@ moverEnemigo enemigoIni delta mapObstaculos =
         posResueltaX = posEnemigo enemyResueltoX
         posY = posResueltaX + V2 0 dy
         enemyY = enemyResueltoX { posEnemigo = posY }
+
+        enemyFin = if checkColision enemyY mapObstaculos
+                      then enemyResueltoX
+                      else enemyY
+
+        nuevoVelGolpe = if esEmpujado
+                        then velGolpeE enemigoIni ^* friccion
+                        else V2 0 0
         
-    in if checkColision enemyY mapObstaculos
-       then enemyResueltoX
-       else enemyY
+    in enemyFin { velGolpeE = nuevoVelGolpe }